+++ webapp/src/components/modals/ObizOrderProcessModal.tsx
@@ -225,7 +225,9 @@ export default function ObizOrderProcessModal(
                   article_references: [
                     { reference: availableArticles[0].reference, quantity: 1 },
                   ],
-                  input_value: amount - (amount * discount) / 100,
+                  input_value: parseFloat(
+                    formatter2Digits.format(amount - (amount * discount) / 100)
+                  ),
                 });
               } else {
                 createTestOrder({